# PERF: speed up from_ragged_array with single-pass cython implementation (Polygon and MultiPolygon)