--- a/shapely/_geometry_helpers.pyx
+++ b/shapely/_geometry_helpers.pyx
@@ -32,6 +32,8 @@ from shapely._pygeos_api cimport (
     import_shapely_c_api,
     PGERR_NAN_COORD,
     PGERR_SUCCESS,
+    PGERR_GEOS_EXCEPTION,
+    PGERR_LINEARRING_NCOORDS,
     PyGEOS_CoordSeq_FromBuffer,
     PyGEOS_CreateGeometry,
     PyGEOS_GetGEOSGeometry,
@@ -56,6 +58,72 @@ def _check_out_array(object out, Py_ssize_t size):
     if out.shape[0] < size:
         raise ValueError(f"out array is too small ({out.shape[0]} < {size})")
     return out
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+cdef int _create_simple_geometry(
+    GEOSContextHandle_t geos_handle,
+    const double[:, :] coord_view,
+    Py_ssize_t idx,
+    unsigned int geom_size,
+    unsigned int dims,
+    int geometry_type,
+    char is_ring,
+    int handle_nan,
+    GEOSGeometry **geom,
+) noexcept nogil:
+    """
+    Helper function to create a geometry with single CoordinateSequence
+    (Point, LineString, LinearRing).
+    """
+    cdef GEOSCoordSequence *seq = NULL
+    cdef int errstate
+    cdef unsigned int actual_geom_size = 0
+
+    errstate = PyGEOS_CoordSeq_FromBuffer(
+        geos_handle, &coord_view[idx, 0], geom_size, dims,
+        is_ring, handle_nan, &seq
+    )
+    if errstate != PGERR_SUCCESS:
+        return errstate
+
+    # check the resulting size to prevent invalid rings
+    if is_ring == 1:
+        if GEOSCoordSeq_getSize_r(geos_handle, seq, &actual_geom_size) == 0:
+            return PGERR_GEOS_EXCEPTION
+        if 0 < actual_geom_size < 4:
+            return PGERR_LINEARRING_NCOORDS
+
+    if geometry_type == 0:
+        geom[0] = GEOSGeom_createPoint_r(geos_handle, seq)
+    elif geometry_type == 1:
+        geom[0] = GEOSGeom_createLineString_r(geos_handle, seq)
+    elif geometry_type == 2:
+        geom[0] = GEOSGeom_createLinearRing_r(geos_handle, seq)
+
+    if geom[0] == NULL:
+        return PGERR_GEOS_EXCEPTION
+
+    return PGERR_SUCCESS
+
+
+def _create_simple_geometry_raise_error(int errstate):
+    """
+    Handle error raising for _create_simple_geometry above (so that the function
+    itself can be fully C function without Python error checking).
+    """
+    if errstate == PGERR_NAN_COORD:
+        raise ValueError(
+            "A NaN, Inf or -Inf coordinate was supplied. Remove the "
+            "coordinate or adapt the 'handle_nan' parameter."
+        )
+    elif errstate == PGERR_LINEARRING_NCOORDS:
+        # the error equals PGERR_LINEARRING_NCOORDS (in shapely/src/geos.h)
+        raise ValueError("A linearring requires at least 4 coordinates.")
+    else:
+        # GEOSException is raised by get_geos_handle
+        return
 
 
 @cython.boundscheck(False)
@@ -122,36 +190,14 @@ def simple_geometries_1d(object coordinates, object indices, int geometry_type, 
                     raise ValueError(
                         f"Index {geom_idx} is missing from the input indices."
                     )
-            errstate = PyGEOS_CoordSeq_FromBuffer(
-                geos_handle, &coord_view[idx, 0], geom_size, dims,
-                is_ring, handle_nan, &seq
+            errstate = _create_simple_geometry(
+                geos_handle, coord_view, idx, geom_size, dims, geometry_type,
+                is_ring, handle_nan, &geom
             )
-            if errstate == PGERR_NAN_COORD:
-                raise ValueError(
-                    "A NaN, Inf or -Inf coordinate was supplied. Remove the "
-                    "coordinate or adapt the 'handle_nan' parameter."
-                )
-            elif errstate != PGERR_SUCCESS:
-                return  # GEOSException is raised by get_geos_handle
-            # check the resulting size to prevent invalid rings
-            if is_ring == 1:
-                if GEOSCoordSeq_getSize_r(geos_handle, seq, &actual_geom_size) == 0:
-                    return  # GEOSException is raised by get_geos_handle
-                if 0 < actual_geom_size < 4:
-                    # the error equals PGERR_LINEARRING_NCOORDS (in shapely/src/geos.h)
-                    raise ValueError("A linearring requires at least 4 coordinates.")
+            if errstate != PGERR_SUCCESS:
+                return _create_simple_geometry_raise_error(errstate)
 
             idx += geom_size
-
-            if geometry_type == 0:
-                geom = GEOSGeom_createPoint_r(geos_handle, seq)
-            elif geometry_type == 1:
-                geom = GEOSGeom_createLineString_r(geos_handle, seq)
-            elif geometry_type == 2:
-                geom = GEOSGeom_createLinearRing_r(geos_handle, seq)
-
-            if geom == NULL:
-                return  # GEOSException is raised by get_geos_handle
 
             out_view[geom_idx] = PyGEOS_CreateGeometry(geom, geos_handle)
 
@@ -237,7 +283,9 @@ def get_parts(object[:] array, bint extract_rings=0):
     return parts, index
 
 
-cdef _deallocate_arr(void* handle, np.intp_t[:] arr, Py_ssize_t last_geom_i):
+@cython.boundscheck(False)
+@cython.wraparound(False)
+cdef void _deallocate_arr(void* handle, np.intp_t[:] arr, Py_ssize_t last_geom_i) noexcept nogil:
     """Deallocate a temporary geometry array to prevent memory leaks"""
     cdef Py_ssize_t i = 0
     cdef GEOSGeometry *g
@@ -390,6 +438,227 @@ def collections_1d(object geometries, object indices, int geometry_type = 7, obj
             geom_idx_1 += collection_size[coll_idx]
 
     return out
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+def _from_ragged_array_polygon(
+    const double[:, ::1] coordinates,
+    const np.int64_t[:] offsets1,
+    const np.int64_t[:] offsets2,
+):
+    """
+    Create Polygons from coordinate and offset arrays.
+    """
+    cdef:
+        Py_ssize_t n_geoms
+        Py_ssize_t i, k
+        Py_ssize_t i1, i2, k1, k2
+        Py_ssize_t n_coords, rings_idx
+        int errstate
+        GEOSContextHandle_t geos_handle
+        GEOSGeometry *ring = NULL
+        GEOSGeometry *geom = NULL
+
+    n_geoms = offsets2.shape[0] - 1
+
+    # A temporary array for the geometries that will be given to CreatePolygon.
+    # For simplicity, we use n_geoms instead of calculating
+    # the max needed size (trading performance for a bit more memory usage)
+    temp_rings = np.empty(shape=(n_geoms, ), dtype=np.intp)
+    cdef np.intp_t[:] temp_rings_view = temp_rings
+    # A temporary array for resulting geometries
+    temp_geoms = np.empty(shape=(n_geoms, ), dtype=np.intp)
+    cdef np.intp_t[:] temp_geoms_view = temp_geoms
+
+    # The final target array
+    result = np.empty(shape=(n_geoms, ), dtype=object)
+    cdef object[:] result_view = result
+
+    cdef unsigned int dims = coordinates.shape[1]
+    if dims not in {2, 3}:
+        raise ValueError("coordinates should N by 2 or N by 3.")
+
+    cdef int ring_type = 2
+    cdef int geometry_type = 6  # MultiPolygon
+    cdef char is_ring = 1
+    cdef int handle_nan = 0
+
+    with get_geos_handle() as geos_handle:
+        with nogil:
+            # iterating through the Polygons
+            for i in range(n_geoms):
+
+                # each part (polygon) can consist of multiple rings
+                # (exterior ring + potentially interior rings(s))
+                i1 = offsets2[i]
+                i2 = offsets2[i + 1]
+
+                # iterating through the rings
+                rings_idx = 0
+                for k in range(i1, i2):
+
+                    # each ring consists of certain number of coords
+                    k1 = offsets1[k]
+                    k2 = offsets1[k + 1]
+                    n_coords = k2 - k1
+                    errstate = _create_simple_geometry(
+                        geos_handle, coordinates, k1, n_coords, dims, ring_type,
+                        is_ring, handle_nan, &ring
+                    )
+                    if errstate != PGERR_SUCCESS:
+                        _deallocate_arr(geos_handle, temp_rings_view, rings_idx)
+                        _deallocate_arr(geos_handle, temp_geoms_view, i - 1)
+                        with gil:
+                            return _create_simple_geometry_raise_error(errstate)
+
+                    temp_rings_view[rings_idx] = <np.intp_t>ring
+                    rings_idx += 1
+
+                if rings_idx > 0:
+                    geom = GEOSGeom_createPolygon_r(
+                        geos_handle,
+                        <GEOSGeometry*> temp_rings_view[0],
+                        <GEOSGeometry**> &temp_rings_view[1 if rings_idx > 1 else 0],
+                        rings_idx - 1
+                    )
+                else:
+                    geom = GEOSGeom_createEmptyPolygon_r(geos_handle)
+                if geom == NULL:
+                    _deallocate_arr(geos_handle, temp_rings_view, rings_idx - 1)
+                    _deallocate_arr(geos_handle, temp_geoms_view, i - 1)
+                    with gil:
+                        return  # GEOSException is raised by get_geos_handle
+
+                temp_geoms_view[i] = <np.intp_t>geom
+
+        for i in range(n_geoms):
+            result_view[i] = PyGEOS_CreateGeometry(<GEOSGeometry *>temp_geoms_view[i], geos_handle)
+
+    return result
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+def _from_ragged_array_multipolygon(
+    const double[:, ::1] coordinates,
+    const np.int64_t[:] offsets1,
+    const np.int64_t[:] offsets2,
+    const np.int64_t[:] offsets3,
+):
+    """
+    Create MultiPolygons from coordinate and offset arrays.
+    """
+    cdef:
+        Py_ssize_t n_geoms
+        Py_ssize_t i, j, k
+        Py_ssize_t i1, i2, j1, j2, k1, k2
+        Py_ssize_t n_coords, rings_idx, parts_idx
+        int errstate
+        GEOSContextHandle_t geos_handle
+        GEOSGeometry *ring = NULL
+        GEOSGeometry *part = NULL
+        GEOSGeometry *geom = NULL
+
+    n_geoms = offsets3.shape[0] - 1
+
+    # A temporary array for the geometries that will be given to CreatePolygon
+    # and CreateCollection. For simplicity, we use n_geoms instead of calculating
+    # the max needed size (trading performance for a bit more memory usage)
+    temp_rings = np.empty(shape=(n_geoms, ), dtype=np.intp)
+    cdef np.intp_t[:] temp_rings_view = temp_rings
+    temp_parts = np.empty(shape=(n_geoms, ), dtype=np.intp)
+    cdef np.intp_t[:] temp_parts_view = temp_parts
+    # A temporary array for resulting geometries
+    temp_geoms = np.empty(shape=(n_geoms, ), dtype=np.intp)
+    cdef np.intp_t[:] temp_geoms_view = temp_geoms
+
+    # The final target array
+    result = np.empty(shape=(n_geoms, ), dtype=object)
+    cdef object[:] result_view = result
+
+    cdef unsigned int dims = coordinates.shape[1]
+    if dims not in {2, 3}:
+        raise ValueError("coordinates should N by 2 or N by 3.")
+
+    cdef int ring_type = 2
+    cdef int geometry_type = 6  # MultiPolygon
+    cdef char is_ring = 1
+    cdef int handle_nan = 0
+
+    with get_geos_handle() as geos_handle:
+        with nogil:
+            # iterating through the MultiPolygons
+            for i in range(n_geoms):
+
+                # getting the indices for their parts
+                i1 = offsets3[i]
+                i2 = offsets3[i + 1]
+
+                # Iterating through the geometry parts
+                parts_idx = 0
+                for j in range(i1, i2):
+
+                    # each part (polygon) can consist of multiple rings
+                    # (exterior ring + potentially interior rings(s))
+                    j1 = offsets2[j]
+                    j2 = offsets2[j + 1]
+
+                    # iterating through the rings
+                    rings_idx = 0
+                    for k in range(j1, j2):
+
+                        # each ring consists of certain number of coords
+                        k1 = offsets1[k]
+                        k2 = offsets1[k + 1]
+                        n_coords = k2 - k1
+                        errstate = _create_simple_geometry(
+                            geos_handle, coordinates, k1, n_coords, dims, ring_type,
+                            is_ring, handle_nan, &ring
+                        )
+                        if errstate != PGERR_SUCCESS:
+                            _deallocate_arr(geos_handle, temp_rings_view, rings_idx)
+                            _deallocate_arr(geos_handle, temp_parts_view, parts_idx - 1)
+                            _deallocate_arr(geos_handle, temp_geoms_view, i - 1)
+                            with gil:
+                                return _create_simple_geometry_raise_error(errstate)
+
+                        temp_rings_view[rings_idx] = <np.intp_t>ring
+                        rings_idx += 1
+
+                    part = GEOSGeom_createPolygon_r(
+                        geos_handle,
+                        <GEOSGeometry*> temp_rings_view[0],
+                        <GEOSGeometry**> &temp_rings_view[1 if rings_idx > 1 else 0],
+                        rings_idx - 1
+                    )
+                    if part == NULL:
+                        _deallocate_arr(geos_handle, temp_rings_view, rings_idx - 1)
+                        _deallocate_arr(geos_handle, temp_parts_view, parts_idx)
+                        _deallocate_arr(geos_handle, temp_geoms_view, i - 1)
+                        with gil:
+                            return  # GEOSException is raised by get_geos_handle
+                    temp_parts_view[parts_idx] = <np.intp_t>part
+                    parts_idx += 1
+
+                geom = GEOSGeom_createCollection_r(
+                    geos_handle,
+                    geometry_type, 
+                    <GEOSGeometry**> &temp_parts_view[0],
+                    parts_idx
+                    )
+                if geom == NULL:
+                    _deallocate_arr(geos_handle, temp_parts_view, parts_idx - 1)
+                    _deallocate_arr(geos_handle, temp_geoms_view, i - 1)
+                    with gil:
+                        return  # GEOSException is raised by get_geos_handle
+
+                temp_geoms_view[i] = <np.intp_t>geom
+
+        for i in range(n_geoms):
+            result_view[i] = PyGEOS_CreateGeometry(<GEOSGeometry *>temp_geoms_view[i], geos_handle)
+
+    return result
 
 
 def _geom_factory(uintptr_t g):