--- a/shapely/tests/test_ragged_array.py
+++ b/shapely/tests/test_ragged_array.py
@@ -399,7 +399,7 @@ def test_mixture_polygon_multipolygon():
     assert_geometries_equal(result, expected)
 
 
-def test_from_ragged_incorrect_rings():
+def test_from_ragged_incorrect_rings_short():
     # too few coordinates for linearring
     coords = np.array([[0, 0], [1, 1]], dtype="float64")
     offsets1 = np.array([0, 2])
@@ -420,8 +420,14 @@ def test_from_ragged_incorrect_rings():
             shapely.GeometryType.POLYGON, coords, (offsets1, offsets2)
         )
 
+
+def test_from_ragged_incorrect_rings_unclosed():
+    # NaNs cause the ring to be unclosed
     coords = np.full((4, 2), np.nan)
     offsets1 = np.array([0, 4])
+    offsets2 = np.array([0, 1])
+    offsets3 = np.array([0, 1])
+
     with pytest.raises(
         shapely.GEOSException,
         match="Points of LinearRing do not form a closed linestring",