This is a follow-up on the original PR adding `from_ragged_array` (https://github.com/shapely/shapely/pull/1559). As mentioned there, I started with a pure python implementation reusing existing building blocks, but for better performance we could later add a custom implementation in cython or C. 
The main benefit is that it can then be done in a single pass, instead of (in case of MultiPolygons) first creating all LinearRings (as python objects), and from that all Polygons, and only then all MultiPolygons (with requiring to clone the geometries at each step).

For `from_ragged_array`, I opted to do this in cython, because we already have similar functionality in cython right now (when passing `indices` to the `shapely.linestrings`, `shapely.polygons` etc functions, then we don't use the C ufunc but the custom implementation in cython).

For now this PR only adds a custom implementation for (Multi)Polygon.

On the MultiPolygons of the nz-building-outlines datasets (3.3 million polygons, small polygons with on average only 8 coordinate pairs), this gives a speedup of around 4.5s to 1.5s (with constructing the same geometries from WKB taking around 2.5s).

For the Watershed Boundaries Dataset (https://www.sciencebase.gov/catalog/item/5a1632b8e4b09fc93dd1720c, less but larger polygons), with this PR that gives around 85ms for from_ragged_array vs 380ms for from_wkb